+++ src/components/Template/templateChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2024-12-05",
+    version: "3.5.0",
+    type: "Bug Fix",
+    affects: ["Styles"],
+    notes: [
+      "Adjusts padding to fix an issue where focus indicators were being cut off in the primary content section.",
+    ],
+  },
+  {
     date: "2024-10-02",
     version: "3.4.0",
     type: "Update",