+++ src/theme/foundations/breakpoints.ts
@@ -1,3 +1,8 @@
+export const headerBreakpoints = {
+  mh: "52em", // "832px",
+  lh: "64em", // "1024px
+};
+
 /**
  * Breakpoints from "src/styles/base/_01-breakpoints.scss"
  * How to use Chakra responsive styles: https://chakra-ui.com/docs/features/responsive-styles
@@ -16,7 +21,11 @@
 export default {
   sm: "20em",
   md: "37.5em",
+  // Temporary and just for the header
+  mh: headerBreakpoints.mh,
   lg: "60em",
+  // Temporary and just for the header
+  lh: headerBreakpoints.lh,
   xl: "80em",
   "2xl": "96em",
 };