+++ CHANGELOG.md
@@ -8,6 +8,33 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 3.5.0 (December 5, 2024)
+
+### Adds
+
+- Adds the `SubNav` component.
+- Adds the `useScrollFadeStyles` hook.
+
+### Updates
+
+- Updates the `MultiSelect` component to add `itemCount` as a data property to the `items` prop to render the item count for an option.
+
+### Fixes
+
+- Fixes issue where focus indicator was being cut off in places in the `Template` component.
+
+## 3.4.4 (November 20, 2024)
+
+### Adds
+
+- Fixes a styling issue in the `Header`'s list to make it stricter.
+
+## 3.4.3 (November 18, 2024)
+
+### Adds
+
+- Adds the NYPL Header as a DS component as a temporary solution for a specific application. It's only meant to be used internally and teams are still encouraged to use the `nypl-header-app`.
+
 ## 3.4.2 (November 7, 2024)
 
 ### Adds